zafferano-archiver/requirements.txt
beautifulsoup4
Flask
Flask-RESTful
psycopg2-binary
requests
